feat: rename files in `i18n` to match kebab-case
